Add setType, getStatusList, getTypeList methods and number of pages and year validators to Blipoteka_Book.
A setType method takes instance of Blipoteka_User and disallows setting a book type from owned to free unless user object passed in is the owner of the book.

A new set of validators restricts year to be in range 1900-current and requires a minimum number of pages to be at least 16 (a previous value of 48 was too restrictive).

Unit tests cover changes described above.